% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM2.R, R/KM2_plots.R, R/KM3.R, R/cox2.R,
%   R/mhazard-deprecated.R
\name{KM2}
\alias{KM2}
\alias{plotKM2.3D}
\alias{plotKM2.HM}
\alias{KM3}
\alias{cox2}
\alias{cox2.LF}
\alias{mhazard-deprecated}
\title{Deprecated functions in package \pkg{mhazard}.}
\usage{
KM2(
  Y1,
  Y2,
  Delta1,
  Delta2,
  newT1 = NULL,
  newT2 = NULL,
  estimator = c("dabrowska", "volterra", "prentice-cai"),
  conf.int = FALSE,
  R = 1000,
  ...
)

plotKM2.3D(
  km2.obj,
  col = "grey",
  shade = 0.25,
  theta = 120,
  xlab = "T1",
  ylab = "T2",
  zlab = "Fhat",
  ...
)

plotKM2.HM(
  km2.obj,
  contour = TRUE,
  col = terrain.colors(100),
  xlab = "T1",
  ylab = "T2",
  ...
)

KM3(
  Y1,
  Y2,
  Y3,
  Delta1,
  Delta2,
  Delta3,
  newT1 = NULL,
  newT2 = NULL,
  newT3 = NULL,
  conf.int = FALSE,
  R = 1000,
  ...
)

cox2(Y1, Y2, Delta1, Delta2, X)

cox2.LF(
  cox2.obj,
  X0_out,
  X1_out,
  T1_out,
  T2_out,
  confidence = c("none", "CI", "CB"),
  n.boot = 100
)
}
\description{
The functions listed below are deprecated and will be defunct in
the near future. Please use the updated version of these functions. Help
pages for deprecated functions are available at
\code{help("<function>-deprecated")}.
}
\section{\code{KM2}}{

For \code{KM2}, use \code{\link{npSurv2}}.
}

\section{\code{plotKM2.3D}}{

For \code{plotKM2.3D}, use \code{\link{plotnpSurv2.3D}}.
}

\section{\code{plotKM2.HM}}{

For \code{plotKM2.HM}, use \code{\link{plotnpSurv2.HM}}.
}

\section{\code{KM3}}{

For \code{KM3}, use \code{\link{npSurv3}}.
}

\section{\code{cox2}}{

For \code{cox2}, use \code{\link{mHR2}}.
}

\section{\code{cox2.LF}}{

For \code{cox2.LF}, use \code{\link{mHR2.LF}}.
}

\keyword{internal}
